Added Manage Listings View etc.

+++ app/Http/Controllers/ListingController.php
@@ -64,6 +64,11 @@ class ListingController extends Controller
     public function update(Request $request, Listing $listing)
     {
 
+        // Make sure logged in user is owner
+        if ($listing->user_id != Auth::id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'company' => ['required'],
             'title' => 'required',
@@ -85,9 +90,21 @@ class ListingController extends Controller
     }
 
     // Delete Listing
-    public function destroy(Listing $listing){
+    public function destroy(Listing $listing)
+    {
+
+        // Make sure logged in user is owner
+        if ($listing->user_id != Auth::id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage Listing
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
